Rename store purchase integration
